Rework chromatic test for transparent select to avoid unnecessary duplication
