+++ packages/react-components/src/select/stories/multiple.stories.mdx
@@ -145,6 +145,21 @@ A multiple values select can display an [icon](?path=/docs/materials-icons--page
     </Story>
 </Preview>
 
+### Disabled
+
+A multiple values select can be disabled.
+
+<Preview>
+    <Story name="disabled">
+        <Select
+            placeholder="Select a galaxy"
+            multiple
+            disabled
+            options={GALAXIES}
+        />
+    </Story>
+</Preview>
+
 ### Fluid
 
 A multiple values select can take all the space available.
@@ -170,22 +185,6 @@ A multiple values select can be transparent.
             placeholder="Select a galaxy"
             multiple
             transparent
-            options={GALAXIES}
-        />
-    </Story>
-</Preview>
-
-
-### Disabled
-
-A multiple values select can be disabled.
-
-<Preview>
-    <Story name="disabled">
-        <Select
-            placeholder="Select a galaxy"
-            multiple
-            disabled
             options={GALAXIES}
         />
     </Story>
