+++ packages/react-components/src/select/tests/chromatic/multiple/multiple.chroma.jsx
@@ -1,0 +1,17 @@
+import { Select } from "@react-components/select";
+import { createChromaticSection, paramsBuilder, storiesOfBuilder } from "@utils";
+import { createSharedStories } from "./shared-stories";
+
+function stories(segment) {
+    return storiesOfBuilder(module, createChromaticSection("Select/multiple"))
+        .segment(segment)
+        .parameters(paramsBuilder()
+            .chromaticDelay(100)
+            .canvasLayout({ width: "80%" })
+            .build())
+        .build();
+}
+
+createSharedStories(<Select />, stories());
+
+createSharedStories(<Select transparent />, stories("/transparent"));
